WiP work to convert to authz manager
