+++ server/springboot/src/main/java/com/styra/tickethub_springboot/dao/model/RequestHeaderInterceptor.java
@@ -28,7 +28,7 @@ public class RequestHeaderInterceptor implements HandlerInterceptor {
         String userName = components[2].trim();
 
         MDC.put("TENANT_NAME", tenantName);
-        MDC.put("USER_NAME", tenantName);
+        MDC.put("USER_NAME", userName);
 
         System.out.printf("DEBUG: setting using MDC tenant='%s' user='%s'\n", tenantName, userName);
