+++ server/springboot/src/main/java/com/styra/tickethub_springboot/web/TicketController.java
@@ -29,70 +29,6 @@ import java.util.Optional;
 @RestController
 public class TicketController {
 
-  private Tenant tenantFromHeader(String authHeader) {
-      var components = authHeader.split("\\s*[\\s+/]\\s*", 3);
-      String tenantName;
-      if (components.length != 3) {
-        System.out.println("ERROR: invalid auth header: " + authHeader);
-        throw new ResponseStatusException(HttpStatus.FORBIDDEN, "auth header is malformed");
-      }
-      tenantName = components[1].trim();
-      Optional<Tenant> tenantBox = tenantRepository.findByName(tenantName);
-      if (!tenantBox.isPresent()) {
-        throw new ResponseStatusException(HttpStatus.FORBIDDEN, "auth header specifies nonexistant tenant");
-      }
-      return tenantBox.get();
-  }
-
-  private String userFromHeader(String authHeader) {
-      var components = authHeader.split("\\s*[\\s+/]\\s*", 3);
-      String user;
-      if (components.length != 3) {
-        System.out.println("ERROR: invalid auth header: " + authHeader);
-        throw new ResponseStatusException(HttpStatus.FORBIDDEN, "auth header is malformed");
-      }
-      user = components[2].trim();
-      return user;
-  }
-
-  private void verifyAuth(String authHeader, String action) {
-      // TODO: opa client object should be re-used
-      String opaURL = "http://localhost:8181";
-      String opaURLEnv = System.getenv("OPA_URL");
-      if (opaURLEnv != null) {
-          opaURL = opaURLEnv;
-      }
-      System.out.printf("DEBUG: using OPA URL: %s\n", opaURL);
-
-      OPAClient opa = new OPAClient(opaURL);
-
-      Tenant tenant = tenantFromHeader(authHeader);
-      String user = userFromHeader(authHeader);
-
-      java.util.Map<String, Object> iMap = java.util.Map.ofEntries(
-          entry("user", user),
-          entry("tenant", tenant.getName()),
-          entry("action", action)
-          );
-
-      System.out.printf("DEBUG: OPA input is: %s\n", iMap);
-
-      boolean allow;
-
-      try {
-        allow = opa.check("tickets/allow", iMap);
-      } catch (Exception e) {
-        System.out.printf("DEBUG: exception while obtaining policy decision: %s\n", e);
-        throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "failed to obtain policy decision");
-      }
-
-      System.out.printf("policy decision was: %s\n", allow);
-
-      if (!allow) {
-        throw new ResponseStatusException(HttpStatus.FORBIDDEN, "access denined by policy");
-      }
-  }
-
   @Autowired
   TicketRepository ticketRepository;
